requirements.txt
# Odoo Integration
odoorpc==0.10.1

# AI/ML
anthropic>=0.25.0
openai>=1.0.0

# Data & File Handling
pydantic>=2.0.0
python-dotenv>=1.0.0

# CLI & UX
click>=8.1.0
rich>=13.0.0
questionary>=2.0.0

# Utilities
python-dateutil>=2.8.0
requests>=2.31.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
